:truck: move `internal/vrl` to `internal/ffi/vrl`
