+++ .dockerignore
@@ -1,7 +1,7 @@
 /bin/
 /data/
 /internal/ffi/filterdsl/rust-crate/target/
-/internal/vrl/rust-crate/target/
+/internal/ffi/vrl/rust-crate/target/
 /web/static/webcomponents/
 /web/components/floweditor/node_modules/
 /web/components/code-editor/node_modules/
