+++ internal/ffi/vrl/errors.go
