+++ internal/ffi/vrl/ffi.go
