+++ internal/ffi/vrl/rust-crate/src/ffi/hmap_struct.rs
