+++ internal/ffi/vrl/rust-crate/src/ffi/result_struct.rs
