+++ internal/ffi/vrl/rust-crate/src/lib.rs
